feat: icons file split
