--- src/components/implements-icon/IconComponent.tsx
@@ -1,181 +1,0 @@
-'use client';
-
-import styled from '@emotion/styled';
-import { useState } from 'react';
-
-import { IconType } from './icon.type';
-
-export function IconComponent({ type }: { type: IconType }) {
-  const [fill, setFill] = useState<string>('#aaa');
-
-  const IconMouseEnter = () => {
-    setFill('#ededed');
-  };
-
-  const IconMouseLeave = () => {
-    setFill('#aaa');
-  };
-
-  return (
-    <styles.container
-      onMouseEnter={IconMouseEnter}
-      onMouseLeave={IconMouseLeave}
-    >
-      {Icon(type, fill)}
-    </styles.container>
-  );
-}
-
-function Icon(type: IconType, fill: string) {
-  switch (type) {
-    case IconType.Arrow:
-      return (
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          width='24'
-          height='25'
-          viewBox='0 0 24 25'
-          fill='none'
-        >
-          <path
-            d='M15.15 21.875C14.7667 22.0583 14.3833 22.0793 14 21.938C13.6167 21.796 13.3333 21.5333 13.15 21.15L10.15 14.7L7.825 17.95C7.54167 18.35 7.16667 18.475 6.7 18.325C6.23333 18.175 6 17.8583 6 17.375V4.55C6 4.13333 6.18733 3.83333 6.562 3.65C6.93733 3.46667 7.29167 3.50833 7.625 3.775L17.725 11.725C18.1083 12.0083 18.2207 12.375 18.062 12.825C17.904 13.275 17.5833 13.5 17.1 13.5H12.9L15.875 19.875C16.0583 20.2583 16.0793 20.6417 15.938 21.025C15.796 21.4083 15.5333 21.6917 15.15 21.875Z'
-            fill={fill}
-          />
-        </svg>
-      );
-    case IconType.Pointer:
-      return (
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          width='24'
-          height='25'
-          viewBox='0 0 24 25'
-          fill='none'
-        >
-          <g clipPath='url(#clip0_29_2)'>
-            <path
-              d='M13 24.5C9.74001 24.5 6.81001 22.51 5.60001 19.48L2.57001 11.87C2.26001 11.08 3.00001 10.29 3.81001 10.55L4.60001 10.81C5.16001 10.99 5.62001 11.42 5.84001 11.97L7.25001 15.5H8.00001V3.75C8.00001 3.06 8.56001 2.5 9.25001 2.5C9.94001 2.5 10.5 3.06 10.5 3.75V12.5H11.5V1.75C11.5 1.06 12.06 0.5 12.75 0.5C13.44 0.5 14 1.06 14 1.75V12.5H15V3.25C15 2.56 15.56 2 16.25 2C16.94 2 17.5 2.56 17.5 3.25V12.5H18.5V6.25C18.5 5.56 19.06 5 19.75 5C20.44 5 21 5.56 21 6.25V16.5C21 20.92 17.42 24.5 13 24.5Z'
-              fill={fill}
-            />
-          </g>
-          <defs>
-            <clipPath id='clip0_29_2'>
-              <rect width='24' height='24' transform='translate(0 0.5)' />
-            </clipPath>
-          </defs>
-        </svg>
-      );
-    case IconType.Table:
-      return (
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          width='24'
-          height='24'
-          viewBox='0 0 475 475'
-          fill='none'
-        >
-          <g>
-            <path
-              d='M461.667,49.963c-8.949-8.947-19.698-13.418-32.265-13.418H45.682c-12.562,0-23.317,4.471-32.264,13.418
-            C4.473,58.912,0,69.663,0,82.228V392.86c0,12.566,4.473,23.309,13.418,32.261c8.947,8.949,19.701,13.415,32.264,13.415h383.72
-            c12.566,0,23.315-4.466,32.265-13.415c8.945-8.952,13.415-19.701,13.415-32.261V82.228
-            C475.082,69.663,470.612,58.909,461.667,49.963z M146.183,392.85c0,2.673-0.859,4.856-2.574,6.571
-            c-1.712,1.711-3.899,2.562-6.567,2.562h-91.36c-2.662,0-4.853-0.852-6.567-2.562c-1.713-1.715-2.568-3.898-2.568-6.571V338.03
-            c0-2.669,0.855-4.853,2.568-6.56c1.714-1.719,3.905-2.574,6.567-2.574h91.363c2.667,0,4.858,0.855,6.567,2.574
-            c1.711,1.707,2.57,3.891,2.57,6.56V392.85z M146.183,283.221c0,2.663-0.859,4.854-2.574,6.564
-            c-1.712,1.714-3.899,2.569-6.567,2.569h-91.36c-2.662,0-4.853-0.855-6.567-2.569c-1.713-1.711-2.568-3.901-2.568-6.564v-54.819
-            c0-2.664,0.855-4.854,2.568-6.567c1.714-1.709,3.905-2.565,6.567-2.565h91.363c2.667,0,4.854,0.855,6.567,2.565
-            c1.711,1.713,2.57,3.903,2.57,6.567V283.221z M146.183,173.587c0,2.666-0.859,4.853-2.574,6.567
-            c-1.712,1.709-3.899,2.568-6.567,2.568h-91.36c-2.662,0-4.853-0.859-6.567-2.568c-1.713-1.715-2.568-3.901-2.568-6.567V118.77
-            c0-2.666,0.855-4.856,2.568-6.567c1.714-1.713,3.905-2.568,6.567-2.568h91.363c2.667,0,4.854,0.855,6.567,2.568
-            c1.711,1.711,2.57,3.901,2.57,6.567V173.587z M292.362,392.85c0,2.673-0.855,4.856-2.563,6.571c-1.711,1.711-3.9,2.562-6.57,2.562
-            H191.86c-2.663,0-4.853-0.852-6.567-2.562c-1.713-1.715-2.568-3.898-2.568-6.571V338.03c0-2.669,0.855-4.853,2.568-6.56
-            c1.714-1.719,3.904-2.574,6.567-2.574h91.365c2.669,0,4.859,0.855,6.57,2.574c1.704,1.707,2.56,3.891,2.56,6.56v54.819H292.362z
-            M292.362,283.221c0,2.663-0.855,4.854-2.563,6.564c-1.711,1.714-3.9,2.569-6.57,2.569H191.86c-2.663,0-4.853-0.855-6.567-2.569
-            c-1.713-1.711-2.568-3.901-2.568-6.564v-54.819c0-2.664,0.855-4.854,2.568-6.567c1.714-1.709,3.904-2.565,6.567-2.565h91.365
-            c2.669,0,4.859,0.855,6.57,2.565c1.704,1.713,2.56,3.903,2.56,6.567v54.819H292.362z M292.362,173.587
-            c0,2.666-0.855,4.853-2.563,6.567c-1.711,1.709-3.9,2.568-6.57,2.568H191.86c-2.663,0-4.853-0.859-6.567-2.568
-            c-1.713-1.715-2.568-3.901-2.568-6.567V118.77c0-2.666,0.855-4.856,2.568-6.567c1.714-1.713,3.904-2.568,6.567-2.568h91.365
-            c2.669,0,4.859,0.855,6.57,2.568c1.704,1.711,2.56,3.901,2.56,6.567v54.817H292.362z M438.536,392.85
-            c0,2.673-0.855,4.856-2.562,6.571c-1.718,1.711-3.908,2.562-6.571,2.562h-91.354c-2.673,0-4.862-0.852-6.57-2.562
-            c-1.711-1.715-2.56-3.898-2.56-6.571V338.03c0-2.669,0.849-4.853,2.56-6.56c1.708-1.719,3.897-2.574,6.57-2.574h91.354
-            c2.663,0,4.854,0.855,6.571,2.574c1.707,1.707,2.562,3.891,2.562,6.56V392.85z M438.536,283.221c0,2.663-0.855,4.854-2.562,6.564
-            c-1.718,1.714-3.908,2.569-6.571,2.569h-91.354c-2.673,0-4.862-0.855-6.57-2.569c-1.711-1.711-2.56-3.901-2.56-6.564v-54.819
-            c0-2.664,0.849-4.854,2.56-6.567c1.708-1.709,3.897-2.565,6.57-2.565h91.354c2.663,0,4.854,0.855,6.571,2.565
-            c1.707,1.713,2.562,3.903,2.562,6.567V283.221z M438.536,173.587c0,2.666-0.855,4.853-2.562,6.567
-            c-1.718,1.709-3.908,2.568-6.571,2.568h-91.354c-2.673,0-4.862-0.859-6.57-2.568c-1.711-1.715-2.56-3.901-2.56-6.567V118.77
-            c0-2.666,0.849-4.856,2.56-6.567c1.708-1.713,3.897-2.568,6.57-2.568h91.354c2.663,0,4.854,0.855,6.571,2.568
-            c1.707,1.711,2.562,3.901,2.562,6.567V173.587z'
-              fill={fill}
-            />
-          </g>
-        </svg>
-      );
-    case IconType.Memo:
-      return (
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          width='24'
-          height='24'
-          viewBox='0 0 24 24'
-          fill='none'
-        >
-          <g id='Layer_Grid' />
-          <g id='Layer_2'>
-            <path
-              d='M9,2C8.4,2,8,2.4,8,3v1.1C6.3,4.6,5,6.1,5,8v3c-1.7,0-3,1.3-3,3s1.3,3,3,3c0,2.2,1.8,4,4,4h7c2.2,0,4-1.8,4-4v-3.1V8   c0-1.9-1.3-3.4-3-3.9V3c0-0.6-0.4-1-1-1s-1,0.4-1,1v1h-5V3C10,2.4,9.6,2,9,2z M6,15H5c-0.6,0-1-0.4-1-1s0.4-1,1-1h1h7   c0.6,0,1,0.4,1,1c0,0.2,0,0.4,0.1,0.6c0,0.1,0,0.1,0,0.2c0,0.1,0,0.2,0.1,0.3H6z M18,17c0,1.1-0.9,2-2,2H9c-1.1,0-2-0.9-2-2h10   c0.1,0,0.2,0,0.3,0c0.3,0,0.5-0.1,0.7-0.2V17z M16,6c1.1,0,2,0.9,2,2v5.9c0,0.6-0.4,1.1-0.9,1.1c-0.1,0-0.2,0-0.3,0   c0,0-0.1,0-0.1,0c-0.1-0.1-0.2-0.1-0.4-0.2C16.1,14.5,16,14.3,16,14c0-1.7-1.3-3-3-3H7V8c0-1.1,0.9-2,2-2H16z'
-              fill={fill}
-            />
-          </g>
-        </svg>
-      );
-    case IconType.IdentifyOneByOne:
-      return (
-        <svg width='24' height='24'>
-          <rect width='24' height='24' fill={fill} />
-        </svg>
-      );
-    case IconType.IdentifyOneByMany:
-      return (
-        <svg width='24' height='24'>
-          <rect width='24' height='24' fill={fill} />
-        </svg>
-      );
-    case IconType.IdentifyManyByMany:
-      return (
-        <svg width='24' height='24'>
-          <rect width='24' height='24' fill={fill} />
-        </svg>
-      );
-    case IconType.NonIdentifyOneByOne:
-      return (
-        <svg width='24' height='24'>
-          <rect width='24' height='24' fill={fill} />
-        </svg>
-      );
-    case IconType.NonIdentifyOneByMany:
-      return (
-        <svg width='24' height='24'>
-          <rect width='24' height='24' fill={fill} />
-        </svg>
-      );
-    case IconType.NonIdentifyManyByMany:
-      return (
-        <svg width='24' height='24'>
-          <rect width='24' height='24' fill={fill} />
-        </svg>
-      );
-    default:
-      break;
-  }
-}
-
-const styles = {
-  container: styled.div`
-    user-select: none;
-    cursor: pointer;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  `,
-};
