+++ src/components/implements-icon/index.ts
@@ -1,2 +1,6 @@
-export * from './IconComponent';
+export * from './RelationshipIcons';
 export * from './icon.type';
+export * from './ArrowIcon';
+export * from './MemoIcon';
+export * from './PointerIcon';
+export * from './TableIcon';
