+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './TableMenuComponent';
-export * from './RelationComponent';
-export * from './ImplementComponent';
-export * from './TableComponent';
+export * from './TableMenu';
+export * from './RelationshipMenu';
+export * from './ErdDrawTools';
+export * from './Table';
